allow a ParseException with just a string
ParseException ctor used to require various objects of the antler (v2)
parse hierarchy, which were useful for debugging. We will probably want
to extend it with antlr4 capabilities. But in the case of sql syntax
not supported by qserv we want to allow ParseError with no reference
to antlr - it just needs to say why the ParseError was thrown.