Track used SELECT ValueExprs in the QueryContext
Stop keeping track of aliases in the QueryContext
Remove the data structure that was used to keep track of aliases.

I decided against having a custom data structure to keep track of the
ValueExprs used in the SELECT list and instead added a vector, a setter,
and a getter to QueryContext. If the complexity of this job increases
much/any it may be prudent to move this code back to a dedicated class.
But for now the current approach seems less cluttered.