add SqlException, for no such database & table
SqlConnection previously reported all pass/fail by bool return value,
which is insufficient in cases where the failure reason needed to be
known. It also was unhelpful to return by bool when the failure needed
to be passed back several frames up the call stack.

Adding the exceptions allows for these things to be handled more
elegantly in the correct location. For consistency it would be good to
change the whole class to be exception-based for error handling, but
that's going to be too much of a change (and work) for these changes.