throw if TableRef::isSubset used with joined TR
The logic for isSubset of a TableRef would have to be designed and is
not needed in the current code. If it does become needed, it might be
worth reflecting on the current abstraction where TableRef owns the
JoinRef, and possibly invert that so the JoinRef owns 2 (or more)
TableRefs and/or a TableRef and another JoinRef(?), but anyway keep
TableRef simple and contain the complexity of Joins entirely in a
separate class.